[CHG] Update namespace of shared
Use <see> to refer .NET components
